skeleton of client now set up
